CU-85zruqf23 - Implement Issue Creation Modal
