+++ frontend/src/app/features/feature-issues/components/issues-filter/issues-filter.component.html
@@ -3,103 +3,11 @@
         <h3 class="col"><strong>Issues</strong></h3>
         <!-- Caption to add -->
         <!-- Button trigger modal -->
-        <button type="button" class="col-1 btn btn-secondary rounded-3" data-bs-toggle="modal" data-bs-target="#exampleModal">
+        <button type="button" class="col-1 btn btn-secondary rounded-3" data-bs-toggle="modal" data-bs-target="#issueModal">
             Create Issue
         </button>
     
-        <!-- Modal -->
-        <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-            <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Create Issue</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                
-                    <form [formGroup]="issueCreationForm" (ngSubmit)="createIssue()">
-                        <div class="modal-body">
-                            <div *ngIf="error" class="alert alert-danger alert-dismissible mt-1 mb-3">
-                                {{error.error.message}}
-                                <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-                            </div>
-
-                            <div class="form-group mt-2">
-                                <label for="issueId">Issue ID</label>
-                                <input type="text" formControlName="issueId" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['issueId'].errors }" />
-                                <div *ngIf="submitted && form['issueId'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['issueId'].errors['required']">Issue ID is required</div>
-                                </div>
-                            </div>
-                            
-                            <div class="form-group mt-2">
-                                <label for="title">Title</label>
-                                <input type="text" formControlName="title" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['title'].errors }" />
-                                <div *ngIf="submitted && form['title'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['title'].errors['required']">Title is required</div>
-                                </div>
-                            </div>
-                            
-                            <div class="form-group mt-2">
-                                <label for="description">Description</label>
-                                <input type="text" formControlName="description" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['description'].errors }" />
-                                <div *ngIf="submitted && form['description'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['description'].errors['required']">Description is required</div>
-                                </div>
-                            </div>
-                            
-                            <div class="form-group mt-2">
-                                <label for="reproducingSteps">Reproducing Steps</label>
-                                <input type="text" formControlName="reproducingSteps" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['reproducingSteps'].errors }" />
-                                <div *ngIf="submitted && form['reproducingSteps'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['reproducingSteps'].errors['required']">Reproducing Steps are required</div>
-                                </div>
-                            </div>
-            
-                            <div class="form-group mt-2">
-                                <label for="environment">Environment</label>
-                                <input type="text" formControlName="environment" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['environment'].errors }" />
-                                <div *ngIf="submitted && form['environment'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['environment'].errors['required']">Environment is required</div>
-                                </div>
-                            </div>
-
-                            <div class="form-group mt-2">
-                                <label for="version">Version</label>
-                                <input type="text" formControlName="version" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['version'].errors }" />
-                                <div *ngIf="submitted && form['version'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['version'].errors['required']">Version is required</div>
-                                </div>
-                            </div>
-
-                            <div class="form-group mt-2">
-                                <label for="priority">Priority</label>
-                                <input type="text" formControlName="priority" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['priority'].errors }" />
-                                <div *ngIf="submitted && form['priority'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['priority'].errors['required']">Priority is required</div>
-                                </div>
-                            </div>
-
-                            <div class="form-group mt-2">
-                                <label for="project">Project</label>
-                                <input type="text" formControlName="project" class="form-control" [ngClass]="{ 'is-invalid': submitted && form['project'].errors }" />
-                                <div *ngIf="submitted && form['project'].errors" class="invalid-feedback">
-                                    <div *ngIf="form['project'].errors['required']">Project is required</div>
-                                </div>
-                            </div>
-                        </div>
-        
-                        <div class="modal-footer">
-                            <button type="button" id="createIssueForm" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                            
-                            <button [disabled]="loading" class="btn btn-primary">
-                                <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
-                                Create
-                            </button>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </div>
+        <app-issue-modal></app-issue-modal>
     </div>
 
     <div class="card rounded-4">
