+++ frontend/src/app/features/feature-issues/components/issues-filter/issues-filter.component.ts
@@ -23,49 +23,49 @@ export class IssuesFilterComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
-    this.issueCreationForm = this.formBuilder.group({
-      issueId: ['', Validators.required],
-      title: ['', Validators.required],
-      description: ['', Validators.required],
-      reproducingSteps: ['', Validators.required],
-      environment: ['', Validators.required],
-      version: ['', Validators.required],
-      priority: ['', Validators.required],
-      project: ['', Validators.required]
-    });
+    // this.issueCreationForm = this.formBuilder.group({
+    //   issueId: ['', Validators.required],
+    //   title: ['', Validators.required],
+    //   description: ['', Validators.required],
+    //   reproducingSteps: ['', Validators.required],
+    //   environment: ['', Validators.required],
+    //   version: ['', Validators.required],
+    //   priority: ['', Validators.required],
+    //   project: ['', Validators.required]
+    // });
   }
 
-  get form() {
-    return this.issueCreationForm.controls;
-  }
+  // get form() {
+  //   return this.issueCreationForm.controls;
+  // }
 
-  createIssue(): void {
-    this.submitted = true;
+  // createIssue(): void {
+  //   this.submitted = true;
 
-    if(!this.issueCreationForm.valid) {
-      return;
-    }
+  //   if(!this.issueCreationForm.valid) {
+  //     return;
+  //   }
 
-    this.loading = true;
+  //   this.loading = true;
 
-    this.issueService.createIssue({
-      issueId: this.form['issueId'].value,
-      title: this.form['title'].value,
-      description: this.form['description'].value,
-      reproducingSteps: this.form['reproducingSteps'].value,
-      environment: this.form['environment'].value,
-      version: this.form['version'].value,
-      priority: this.form['priority'].value,
-    }, this.form['project'].value).subscribe({
-      next: () => {
-        document.getElementById('createIssueForm')?.click();
-        window.location.reload();
-      },
-      error: error => {
-        this.error = error;
-        this.loading = false;
-      }
-    });
-  }
+  //   this.issueService.createIssue({
+  //     issueId: this.form['issueId'].value,
+  //     title: this.form['title'].value,
+  //     description: this.form['description'].value,
+  //     reproducingSteps: this.form['reproducingSteps'].value,
+  //     environment: this.form['environment'].value,
+  //     version: this.form['version'].value,
+  //     priority: this.form['priority'].value,
+  //   }, this.form['project'].value).subscribe({
+  //     next: () => {
+  //       document.getElementById('createIssueForm')?.click();
+  //       window.location.reload();
+  //     },
+  //     error: error => {
+  //       this.error = error;
+  //       this.loading = false;
+  //     }
+  //   });
+  // }
 
 }
